CRUD Commit

+++ app/Http/Controllers/UserController.php
@@ -3,10 +3,15 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use App\Models\User;
+use App\Traits\ApiResponser;
+
 
 class UserController extends Controller
 {
+
+    use ApiResponser;
     private $request;
 
     public function __construct(Request $request){
@@ -28,18 +33,23 @@ class UserController extends Controller
         $rules = [
         'username' => 'required|max:20',
         'password' => 'required|max:20',
-        'gender' => 'required|in:Male,Female',
         ];
+
         $this->validate($request,$rules);
         $user = User::create($request->all());
-        return $this->successResponse($user,
-        Response::HTTP_CREATED);
+        return $this->successResponse($user, Response::HTTP_CREATED);
     }
 
     public function show($id)
     {
-        $user = User::findOrFail($id);
-        return $this->successResponse($user);
+        $user = User::find($id);
+        // $user = User::where('id', $id)->first();
+        if($user){
+            return $this->successResponse($user);
+        }
+        else{
+            return $this->errorResponse('User ID is not found', Response::HTTP_NOT_FOUND);
+        }
     }
 
 
@@ -48,28 +58,38 @@ class UserController extends Controller
         $rules = [
         'username' => 'max:20',
         'password' => 'max:20',
-        'gender' => 'in:Male,Female',
         ];
 
         $this->validate($request, $rules);
-        $user = User::findOrFail($id);
+        $user = User::find($id);  
 
-        $user->fill($request->all());
-        // if no changes happen
-        if ($user->isClean()) {
-            return $this->errorResponse('At least one value must change', Response::HTTP_UNPROCESSABLE_ENTITY);
+
+        if($user){
+            $user->fill($request->all());
+            // if no changes happen
+            if ($user->isClean()) {
+                return $this->errorResponse('At least one value must change', Response::HTTP_UNPROCESSABLE_ENTITY);
+            }
+
+            $user->save();
+            return $this->successResponse($user);
         }
-
-        $user->save();
-        return $this->successResponse($user);
+        else{
+            return $this->errorResponse('User ID is not found', Response::HTTP_NOT_FOUND);
+        }
     }
 
     public function delete($id)
     {
-        $user = User::findOrFail($id);
-        $user->delete();
+        $user = User::find($id);
+        if($user){
+            $user->delete();
 
-        return $this->successResponse($user);
+            return $this->successResponse($user);
+        }
+        else{
+            return $this->errorResponse('User ID is not found', Response::HTTP_NOT_FOUND);     
+        }
     }
     
 }
